refactor(runtime): split UI state and action modules
